FlaskServer/pyproject.toml
[tool.poetry]
name = "flaskserver"
version = "0.0.0"
description = "smarterhome python"
authors = ["rtimmons"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
cryptography = "^2.9.2"
pyyaml = "^5.3.1"
pyeight = "^0.1.4"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
